requirements.txt
Flask==2.3.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Werkzeug==2.3.7
python-dotenv==1.0.0
Flask-SQLAlchemy==3.0.5
bcrypt==4.0.1
requests==2.31.0
nltk==3.8.1
numpy==1.24.3
scikit-learn==1.2.2
python-docx==1.1.2
PyPDF2==3.0.1
schedule==1.2.0
apscheduler==3.10.1
pytest==7.4.0
